app/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.4
lcu-driver==3.0.0
multidict==6.0.4
mysql-connector-python==8.2.0
protobuf==4.21.12
psutil==5.9.6
requests==2.31.0
riotwatcher==3.2.5
urllib3==2.1.0
yarl==1.9.2
